+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 - `Innmind\OperatingSystem\CurrentProcess\Signals` is now a final class
 - `Innmind\OperatingSystem\CurrentProcess` is now a final class
 - `Innmind\OperatingSystem\OperatingSystem::map()` callable must now return a `Config`
+- `Innmind\OperatingSystem\Config::of()` has been renamed `::new()`
 
 ### Fixed